auth_service/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
PyJWT>=2.8.0
cryptography>=41.0.0
pyyaml>=6.0.1

